Split out database portion from `BeatmapManager`
